firebaseの認証のエラーを解決
