+++ program/app.js
@@ -1,11 +1,31 @@
 const express = require("express");
+require("dotenv").config({ debug: true });
+const bodyParser = require('body-parser');
 const path = require('path');
 let app = express();
 const pg = require("pg");
 
+//const auth = require("./firebase");
 
-//API_KEY
-require("dotenv").config({ debug: true });
+const firebase = require('firebase/compat/app');
+require('firebase/compat/auth');
+
+const firebaseConfig = {
+  apiKey: process.env.API_KEY,
+  authDomain: process.env.AUTH_DOMAIN,
+  projectId: process.env.PROJECT_ID,
+  storageBucket: process.env.STORAGE_BUCKET,
+  messagingSenderId: process.env.MESSAGING_SENDER_ID,
+  appId: process.env.APP_ID,
+};
+
+//console.log(firebaseConfig);
+
+const initialize = firebase.initializeApp(firebaseConfig);
+const auth = initialize.auth();
+
+
+//API_KEY & DB
 
 //API_KEY error check
 if (typeof process.env.API_KEY === "undefined") {
@@ -25,38 +45,49 @@ var pool = new pg.Pool({
 
 
 app.use(express.static("public"));
+app.use(bodyParser.json());
 
 app.get("/", (req, res) => {
     res.sendFile(path.join(__dirname, 'public', 'mainpage', 'main.html'));
 });
 
+
+//firebase認証
+
 app.post("/signup", async(req, res) => {
-  const getuserEmail = req.body.email;
-  console.log("email:", getuserEmail);
-  const getuserName = req.body.displayName;
-  const getuserID = req.uid;
- try {
 
-  const client = await pool.connect();
-  await client.query ("INSERT INTO users (user_name, email, firebase_id) VALUES ($1, $2, $3)", [getuserName, getuserEmail, getuserID]);
-  client.release();
+  const catchEmail = req.body.email;
+  const catchPassword = req.body.password;
+
+  const firebaseResult = await firebase.auth().createUserWithEmailAndPassword(catchEmail, catchPassword)
+      .catch(function(error) {
+        console.log('ログインできません（' + error.message + '）');
+      });
+  console.log(firebaseResult);
+
+  const getuserEmail = firebaseResult.user.email;
+  const getuserName = firebaseResult.user.email;
+  const getuserID = firebaseResult.user.uid;
+
+  try {
+
+    const client = await pool.connect();
+    await client.query ("INSERT INTO users (user_name, email, firebase_id) VALUES ($1, $2, $3)", [getuserName, getuserEmail, getuserID]);
+    client.release();
 
 
-  console.log("post動いてる？")
-  const redirectpage = 'main2.html';
-  res.redirect('/public/mainpage/' + redirectpage);
- } catch (err) {
-  console.log(err);
-  res.status(500).send("データベースエラーが発生しました");
- }
+    console.log("post動いてる？")
+    const redirectpage = 'main2.html';
+    res.redirect('/public/mainpage/' + redirectpage);
+  } catch (err) {
+    console.log(err);
+    res.status(500).send("データベースエラーが発生しました");
+  }
 
 
 })
 
-// app.get("/signup", (req, res) => {
-//   const redirectpage = "/main2.html";
-//   res.redirect(redirectpage);
-// })
+
 
 pool.connect();
