+++ program/firebase.js
@@ -1,15 +1,21 @@
-import firebase from 'firebase/compat/app';
-import 'firebase/compat/auth';
+// import firebase from 'firebase/compat/app';
+// import 'firebase/compat/auth';
 
-const firebaseConfig = {
-  apiKey: process.env.API_KEY,
-  authDomain: process.env.AUTH_DOMAIN,
-  projectId: process.env.PROJECT_ID,
-  storageBucket: process.env.STORAGE_BUCKET,
-  messagingSenderId: process.env.MESSAGING_SENDER_ID,
-  appId: process.env.APP_ID
-};
+// // const firebase = require('firebase/compat/app');
+// // require('firebase/compat/auth');
 
-firebase.initializeApp(firebaseConfig);
+// const firebaseConfig = {
+//   apiKey: process.env.API_KEY,
+//   authDomain: process.env.AUTH_DOMAIN,
+//   projectId: process.env.PROJECT_ID,
+//   storageBucket: process.env.STORAGE_BUCKET,
+//   messagingSenderId: process.env.MESSAGING_SENDER_ID,
+//   appId: process.env.APP_ID
+// };
 
-export const auth = firebase.auth();
+// firebase.initializeApp(firebaseConfig);
+// // // const auth = firebase.auth();
+
+// export const auth = firebase.auth();
+//export default firebase;
+//module.exports = auth;
